pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.14",
    "google-api-python-client>=2.174.0",
    "google-auth>=2.40.3",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "langgraph>=0.5.0",
    "langgraph-checkpoint>=2.1.0",
    "openai>=1.93.0",
    "pydantic>=2.11.7",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.32.4",
    "streamlit>=1.46.1",
    "uvicorn>=0.35.0",
]
